server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
    "websockets>=15.0.1",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.0"
]
